+++ frontend/src/game.ts
@@ -221,8 +221,9 @@ export default class Game extends Screen {
         }
     }
 
+    // Handles key press events, whose key codes are sent down by the application layer
     handleKeyPress = (keyCode: number) => {
-        console.log(keyCode);
+        if (this._engine.currentView) this._engine.handleKeyPress(keyCode);
     }
 
     // Resets all game parameters; copied from the constructor function